Move new utilities into D.B.Internal.Utils
