+++ Data/ByteString.hs
@@ -231,6 +231,7 @@ import Prelude hiding           (reverse,head,tail,last,init,null
 import Data.Bits                (finiteBitSize, shiftL, (.|.), (.&.))
 
 import Data.ByteString.Internal
+import Data.ByteString.Internal.Utils
 import Data.ByteString.Lazy.Internal (fromStrict, toStrict)
 import Data.ByteString.Unsafe
