+++ Data/ByteString/Lazy/Internal.hs
@@ -50,6 +50,7 @@ module Data.ByteString.Lazy.Internal (
 import Prelude hiding (concat)
 
 import qualified Data.ByteString.Internal as S
+import qualified Data.ByteString.Internal.Utils as S
 
 import Data.Word (Word8)
 import Foreign.Storable (Storable(sizeOf))
